requirements.txt
python-dotenv==1.1.0
openai==1.81.0
boto3==1.38.21
pydantic==2.11.4
twine==6.1.0
wheel
setuptools
